+++ packages/headless/src/api/index.ts
@@ -1,3 +1,3 @@
 export * from './services';
 export * from './apolloClient';
-export { ContentNodeOptions, ListPostOptions } from './queries';
+export { ContentNodeOptions, ListPostOptions, categoryOptions } from './queries';
