+++ packages/headless/src/api/queries/index.ts
@@ -1,5 +1,6 @@
 export * from './getPosts';
 export * from './getContentNode';
+export * from './categoryOptions';
 export * from './GENERAL_SETTINGS';
 export * from './GET_URI_INFO';
 export * from './PAGE_DATA_FRAGMENT';
